资金流水 资金费用适配器及时间选择器逻辑的修改

+++ base/src/main/java/com/black/base/model/wallet/Order.kt
@@ -3,7 +3,7 @@ package com.black.base.model.wallet
 import com.black.base.model.BaseAdapterItem
 
 class Order() : BaseAdapterItem()  {
-    var startTime: Int? = null //成交均价
-    var endTime: Int? = null
+    var startTime: Long? = null //成交均价
+    var endTime: Long? = null
 
 }
